chore/BB-19 implements normalization

+++ greenhouse/src/main/java/com/bluebell/platform/models/core/nonentities/portfolio/PortfolioAccountEquityPoint.java
@@ -33,5 +33,5 @@ public class PortfolioAccountEquityPoint {
     @Getter
     @Setter
     @Schema(description = "Normalized account value")
-    private int normalized;
+    private double normalized;
 }
